refactor: separated root component
