+++ package.json
@@ -23,6 +23,7 @@
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.2.0",
+    "react-grid-layout": "^1.3.4",
     "react-hook-form": "^7.43.9",
     "react-intl": "^6.4.2",
     "react-redux": "^8.0.5",
