+++ src/modules/todo/components/todo-list/TodoList.tsx
@@ -1,0 +1,13 @@
+import { TodoItem } from '@/modules/todo/types/todo.types'
+
+type TodoListProps = {
+    items: TodoItem[]
+}
+
+export default function TodoList({ items }: TodoListProps) {
+    return <ul>
+        {items.map((todo) => {
+            return <li key={todo.id}>{todo.text}</li>
+        })}
+    </ul>
+}
