+++ src/modules/todo/components/todo/Todo.tsx
@@ -1,18 +1,13 @@
-import { useForm } from 'react-hook-form'
 import * as classes from './Todo.module.scss'
 import { useCreateTodo } from '@/modules/todo/hooks/useMutationToPostTodo'
 import { useFetchTodos } from '@/modules/todo/hooks/useFetchTodos'
-
-type AddTodoFormValues = {
-  text: string
-}
+import TodoList from '../todo-list/TodoList'
+import AddTodoForm from '../add-todo-form/AddTodoForm'
 
 function Todo() {
-  const { register, handleSubmit, reset } = useForm<AddTodoFormValues>()
-
   const todoList = useFetchTodos()
 
-  const {mutate: todoCreate} = useCreateTodo()
+  const { mutate: todoCreate } = useCreateTodo()
 
   const onSubmit = handleSubmit(({ text }) => {
     todoCreate({
@@ -25,20 +20,9 @@ function Todo() {
   return (
     <div>
       <h1 className={classes.root}>Todo list</h1>
-      <ul>
-        {todoList.isLoading && 'loading...'}
-        {todoList.data?.map((todo) => {
-          return <li key={todo.id}>{todo.text}</li>
-        })}
-      </ul>
-      <form onSubmit={onSubmit}>
-        <label>
-          Add New To-Do
-          <br />
-          <input type="text" {...register('text')} />
-        </label>
-        <input type="submit" value="Add To-Do" />
-      </form>
+      {todoList.isLoading && 'loading...'}
+      {todoList.data && <TodoList items={todoList.data} />}
+      <AddTodoForm />
     </div>
   )
 }
